+++ src/iso/metadata/index.ts
@@ -7,7 +7,7 @@ import {
 } from './replacements';
 import { processLayoutMirror, buildMirrors } from './mirrors';
 import { saveFullSceneModel } from './models';
-import { loadGrid } from '../grid';
+import { getGridMetadata } from '../grid';
 import { loadImageData, loadBricks } from '..';
 import { getPalette, getGrids } from '../../resources';
 import { checkVariantMatch } from './matchers/variants';
@@ -34,13 +34,14 @@ export async function extractGridMetadata(grid, entry, ambience, is3D, numActors
 }
 
 export async function saveSceneReplacementModel(entry, ambience) {
-    const [palette, bkg, bricks, mask] = await Promise.all([
+    const [palette, bricks, gridMetadata, mask] = await Promise.all([
         getPalette(),
-        getGrids(),
         loadBricks(),
+        getGridMetadata(entry + 1),
         loadImageData('images/brick_mask.png')
     ]);
-    const grid = await loadGrid(bkg, bricks, mask, palette, entry + 1, true);
+
+    const grid = await getGrids(entry + 1, { bricks, mask, palette, is3D: true, gridMetadata });
 
     const metadata = await loadMetadata(entry, grid.library, true);
     const replacements = await initReplacements(entry, metadata, ambience, 0);
