+++ src/resources/parse.ts
@@ -9,6 +9,7 @@ import { parseScene } from './parsers/scene';
 import { parseSpriteClipInfo, parseSprite, parseSpriteRaw } from './parsers/sprite';
 import { parseSceneMap, parseBrick } from './parsers/bricks';
 import { parseLibrary } from './parsers/libraries';
+import { parseGrid } from './parsers/grids';
 
 const NOP = (resource: Resource) => {
     return resource;
@@ -35,7 +36,7 @@ const ResourceTypes = {
     BL2: { type: 'BL2', description: 'LBA2 Layout Library', parser: parseLibrary },
     BRK: { type: 'BRK', description: 'LBA Brick Sprite', parser: parseBrick },
     GR1: { type: 'GR1', description: 'LBA1 Grids', parser: NOP },
-    GR2: { type: 'GR2', description: 'LBA2 Grids', parser: NOP },
+    GR2: { type: 'GR2', description: 'LBA2 Grids', parser: parseGrid },
     '3DE': { type: '3DE', description: 'LBA1 File 3D Entity', parser: NOP },
     F3D: { type: 'F3D', description: 'LBA2 Entity Information', parser: parseEntity },
     PAL: { type: 'PAL', description: 'LBA Palette', parser: parsePalette },
